Okay, they all pass on Windows, lets see how Unix is doing

+++ spec.rb
@@ -4,21 +4,25 @@ require_relative 'run'
 RSpec.describe 'the process' do
   def assert_dead(pids)
     Array(pids).each do |pid|
-      message = <<-MESSAGE.gsub(/^ */, '')
-        Expected Errno::ESRCH when checking pid #{pid.inspect},
-        Which means that it can't find the process (b/c its dead)
-        This was not raised, which means you should have orphans
-        So check your processes to make sure. If you have them, kill them by hand,
-        if not, this test needs to be improved to actually check for orphans.
-      MESSAGE
-      expect { Process.kill 0, pid }.to raise_error(Errno::ESRCH), message
+      if ChildProcess.windows?
+        ChildProcess::Windows::Lib.alive? pid
+      else
+        message = <<-MESSAGE.gsub(/^ */, '')
+          Expected Errno::ESRCH when checking pid #{pid.inspect},
+          Which means that it can't find the process (b/c its dead)
+          This was not raised, which means you should have orphans
+          So check your processes to make sure. If you have them, kill them by hand,
+          if not, this test needs to be improved to actually check for orphans.
+        MESSAGE
+        expect { Process.kill 0, pid }.to raise_error(Errno::ESRCH), message
+      end
     end
   end
 
   it 'can echo stdin to stdout' do
     result = run stdin: "abc", program: 'ruby', argv: ['-e', '$stdout.puts $stdin.gets']
     expect(result.code).to eq 0
-    expect(result.stdout).to eq "abc\n"
+    expect(result.stdout.chomp).to eq "abc"
     expect(result.stderr).to eq ""
   end
 
@@ -26,7 +30,7 @@ RSpec.describe 'the process' do
     result = run stdin: "abc", program: 'ruby', argv: ['-e', '$stderr.puts $stdin.gets']
     expect(result.code).to eq 0
     expect(result.stdout).to eq ""
-    expect(result.stderr).to eq "abc\n"
+    expect(result.stderr.chomp).to eq "abc"
   end
 
   it 'times out on stuck processes' do
@@ -41,19 +45,21 @@ RSpec.describe 'the process' do
     start_time = Time.now
     result = run stdin: "abc", program: 'ruby', argv: ['-e', '
       puts $$
-      spawn "ruby", "-e", "puts $$; $stdout.flush; sleep"
+      pid = spawn %(ruby), %(-e), %(sleep)
+      puts pid
+      $stdout.flush
       sleep
     '], timeout: 1
     end_time = Time.now
 
     # it did timeout
-    expect(end_time - start_time).to be > 1
     expect(result.code).to_not eq 0
     expect(result.code).to be_a_kind_of Integer
+    expect(end_time - start_time).to be > 1
 
     # the two children printed their pids
     pids = result.stdout.lines.map do |line|
-      expect(line).to match /^\d+$/
+      expect(line.chomp).to match /^\d+\r?\n?$/
       line.to_i
     end
     expect(pids.length).to eq 2
@@ -64,11 +70,16 @@ RSpec.describe 'the process' do
   it 'cleans up the process and all its children when the parent is interrupted' do
     read, write = IO.pipe
     filepath = File.realpath('run', __dir__)
-    program_pid = spawn filepath, 'ruby', '-e', '
+    program = ChildProcess.build 'ruby', filepath, '--', 'ruby', '-e', '
       puts $$
-      spawn "ruby", "-e", "puts $$; $stdout.flush; sleep"
+      pid = spawn %(ruby), %(-e), %(sleep)
+      puts pid
+      $stdout.flush
       sleep
-    ', out: write, err: write
+    '
+    program.io.stdout = write
+    program.io.stderr = write
+    program.start
     write.close
 
     # get the pids
@@ -78,10 +89,9 @@ RSpec.describe 'the process' do
     expect(grandchild_pid).to match /^\d+$/
 
     # interrupt the program
-    Process.kill 'INT', program_pid
-    Process.wait program_pid
+    program.stop
 
-    assert_dead [program_pid, child_pid.to_i, grandchild_pid.to_i]
+    assert_dead [program.pid, child_pid.to_i, grandchild_pid.to_i]
   end
 
   xit 'cleans up the process and all its children when the child exits normally' do
